% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_zone_description.R
\name{add_zone_description}
\alias{add_zone_description}
\title{Create a data.frame to show the zones}
\usage{
add_zone_description(
  dt,
  zone_dt,
  start_id_col,
  end_id_col,
  zone_id_col,
  keep_dt_id_cols = FALSE
)
}
\arguments{
\item{dt}{Main data.table to be use.}

\item{zone_dt}{A data.table with zones from data dictionary.}

\item{start_id_col}{A string indicating the id to be used as starting point.}

\item{end_id_col}{A string indicating the id to be used as ending point.}

\item{zone_id_col}{A string indicating the id related to each location in the \code{zone_dt} table.}

\item{keep_dt_id_cols}{If \code{FALSE}, the columns indicated the \code{start_id_col} and \code{end_id_col} will be removed.}
}
\value{
A data.table
}
\description{
It translate the start and end ids with more detailed information.
}
\examples{
set.seed(123)

base_data <- data.table::data.table(
  start_id = 1:5,
  end_id = sample.int(5, 5),
  value = seq(50, length.out = 5, by = 50)
)

zone_description <- data.table::data.table(
  id = 1:5,
  City = LETTERS[1:5],
  Zone = tail(LETTERS, 5L)
)

add_zone_description(base_data,
                     zone_description,
                     start_id_col = "start_id",
                     end_id_col = "end_id",
                     zone_id_col = "id")
}
