% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_base_cleaning.R
\name{apply_base_cleaning}
\alias{apply_base_cleaning}
\title{Apply the transformations based on initial EDA}
\usage{
apply_base_cleaning(dt)
}
\arguments{
\item{dt}{A data.table with all columns of data and after translating the code zone with their meaning.}
}
\value{
A data.table
}
\description{
Apply the transformations based on initial EDA
}
\examples{

data.table::data.table(hvfhs_license_num = "HV0004",
                       dispatching_base_num = "HOLA",
                       originating_base_num = NA_character_,
                       access_a_ride_flag = "",     
                       shared_request_flag = "Y",
                       shared_match_flag = "Y",
                       wav_request_flag = "Y",
                       wav_match_flag = "Y",
                       start_borough = "Manhattan",
                       end_borough = "Brooklyn",
                       start_service_zone = "Boro Zone",
                       end_service_zone = "Yellow Zone",
                       start_zone = "1",
                       end_zone = "2",
                       base_passenger_fare = 800,
                       trip_miles = 250,
                       request_datetime = as.POSIXct("2022-05-18 7:25:00"), 
                       dropoff_datetime = as.POSIXct("2022-05-18 8:15:00"),
                       tips = 500,
                       driver_pay = 800) |>
  apply_base_cleaning()

}
