% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bussiness-understanding-baselinesimulation.R
\docType{data}
\name{BaseLineSimulation}
\alias{BaseLineSimulation}
\title{BaseLineSimulation}
\format{
A data frame of 10 variables:
\describe{
\item{ simulation_day }{ Sample date }
\item{ s_PULocationID }{ TLC Taxi Zone in which the trip began }
\item{ s_DOLocationID }{ TLC Taxi Zone in which the trip ended }
\item{ s_request_datetime }{ Date and time when passenger requested to be picked up }
\item{ s_dropoff_datetime }{ The date and time of the trip drop-off }
\item{ s_driver_pay }{ Total driver pay (not including tolls or tips and net of commission, surcharges, or taxes) }
\item{ s_tips }{ Total amount of tips received from passenger }
\item{ PULocationID }{ Start zone of the simulation day }
\item{ hours_to_work }{ The number of hour to work in the simulation day }
\item{ start_time }{ Start date and time for the simulation day }
}
}
\source{
TLC Trip Record Data
}
\usage{
BaseLineSimulation
}
\description{
Trips returned after runing the simulation.
}
\keyword{datasets}
