% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_zone_manually.R
\name{clean_zone_manually}
\alias{clean_zone_manually}
\title{Clean zone errors}
\usage{
clean_zone_manually(dt)
}
\arguments{
\item{dt}{A \code{data.table} with columns \code{lat}, \code{long}, LocationID and Zone.}
}
\value{
A data.table
}
\description{
Clean zone errors
}
