% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_boxplot.R
\name{compute_boxplot}
\alias{compute_boxplot}
\title{Calculates all values need to describe}
\usage{
compute_boxplot(x, value)
}
\arguments{
\item{x}{A data.frame or an arrow connection}

\item{value}{A column vector to summarize}
}
\value{
A data.frame with summary values
}
\description{
Calculates of values needed to create a boxplot.
}
\examples{

data.frame(x = rnorm(200, mean = 5)) |>
  arrow::arrow_table() |>
  compute_boxplot(x)

}
