% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_pct.R
\name{count_pct}
\alias{count_pct}
\title{Counts rows based con columns provided}
\usage{
count_pct(x, ..., wt = NULL, sort = TRUE, digits = 3L)
}
\arguments{
\item{x}{A lazy data frame arrow connection.}

\item{...}{Variables to group by}

\item{wt}{If a variable, computes sum(wt) for each group.}

\item{sort}{If TRUE, will show the largest groups at the top.}

\item{digits}{Integer indicating the number of decimal places}
}
\value{
A data.frame
}
\description{
It counts the number rows where each unique value repeated in the columns
selected arranging there results in descent order and adds a percentage
column after collecting the results from \code{arrow}.
}
\examples{
set.seed(1234)

arrow_con <- 
  data.frame(char1 = sample(LETTERS[1:3], 100, replace = TRUE),
             char2 = sample(LETTERS[4:6], 100, replace = TRUE)) |>
  arrow::arrow_table()

# One var
count_pct(arrow_con, char1)

# Two vars
dt1 <-count_pct(arrow_con, char1, char2)
dt1

# You can use the wt function
count_pct(dt1, char1, wt = n)
}
