% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_weekday.R
\name{factor_weekday}
\alias{factor_weekday}
\title{Transform number into factors}
\usage{
factor_weekday(x)
}
\arguments{
\item{x}{A numeric vector with values from 1 to 7}
}
\value{
A factor vector
}
\description{
Transform a numeric vector in factor with a level for each day of a week.
}
\examples{

factor_weekday(c(1,7,5))

}
