% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_glimpse.R
\name{fast_glimpse}
\alias{fast_glimpse}
\title{glimpse for arrow connections}
\usage{
fast_glimpse(x)
}
\arguments{
\item{x}{an arrow connection}
}
\value{
The schema related to the connection
}
\description{
Shows the class related to each column without loading the data into memory.
}
\examples{

df <- data.frame(num = runif(10),
                 char = LETTERS[1:10],
                 factor = factor(LETTERS[1:10], levels = LETTERS[1:10]),
                 dates = seq(from = as.Date("2024-01-01"), 
                             by = "month", length.out = 10))

arrow::arrow_table(df) |> fast_glimpse()
}
