% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_zones.R
\name{join_zones}
\alias{join_zones}
\title{Create a data.frame to show the zones}
\usage{
join_zones(df, zone_tb)
}
\arguments{
\item{df}{A data.frame with all the code zones from main data.}

\item{zone_tb}{A data.frame with zones from data dictionary.}
}
\value{
A data.table
}
\description{
It creates the \code{start_borough}, \code{start_zone} and \code{start_service_zone}
columns based on the \code{PULocationID} column and the \code{end_borough}, \code{end_zone}
and \code{end_service_zone} columns based on the \code{DOLocationID} column.
}
