% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_cols_to_use.R
\name{list_cols_to_use}
\alias{list_cols_to_use}
\title{Returns a vectors with all the variables use to create the model}
\usage{
list_cols_to_use()
}
\description{
It's a character vector of 14 columns organized in 2 groups:
\enumerate{
\item The first 11 columns are related with data available before arriving at the pick-up location and will be use as predictors.
\item The last 3 columns are related with data available after ending the trip and will be used to estimate calculate the value to predict.
}

You can learnt more about this selection in the \code{vignette("articles/02-business-understanding")}
}
\examples{
list_cols_to_use()
}
