% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_sample_by_group.R
\name{plan_sample_by_group}
\alias{plan_sample_by_group}
\title{Define how to sample by group}
\usage{
plan_sample_by_group(x, ..., size = 1e+06, seed = 1)
}
\arguments{
\item{x}{A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}

\item{...}{Variables to group by.}

\item{size}{Defining the number of row to sample.}

\item{seed}{Define a seed to keep the process reproducible}
}
\value{
A data.table
}
\description{
Define how to sample by group
}
\examples{

data.frame(x = rep(1:5, each = 2)) |>
  plan_sample_by_group(x , size = 5)

}
