% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_initial_cleaning.R
\name{set_initial_cleaning}
\alias{set_initial_cleaning}
\title{Apply the transformations based on initial EDA}
\usage{
set_initial_cleaning(dt)
}
\arguments{
\item{dt}{A data.table with columns:
\itemize{
\item hvfhs_license_num
\item dispatching_base_num
\item originating_base_num
\item access_a_ride_flag
\item request_datetime
\item dropoff_datetime
\item driver_pay
}}
}
\description{
Apply the transformations based on initial EDA
}
