% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm_if_best_trip.R
\name{set_take_current_trip}
\alias{set_take_current_trip}
\title{Confirm whether a taxi driver should take the current trip}
\usage{
set_take_current_trip(
  conn,
  start_points,
  max_min = 15,
  trip_criteria = c("one_trip", "median", "75\%")
)
}
\arguments{
\item{conn}{Duckdb connection to NycTrips and PointMeanDistance tables.}

\item{start_points}{A data.table all trips to evaluate.}

\item{max_min}{Defines the time limit and distance to find better trips}

\item{trip_criteria}{Defines the criteria we want to use when selecting a trip.}
}
\value{
A data.table.
}
\description{
Confirm whether a taxi driver should take the current trip
}
