% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_to_boxplot.R
\name{summary_to_boxplot}
\alias{summary_to_boxplot}
\title{Display boxplot chart based of a summary.}
\usage{
summary_to_boxplot(x, y_title = "", y_break_width = 2)
}
\arguments{
\item{x}{A data.frame}

\item{y_title}{A character value indicating the title to describe the y axis.}

\item{y_break_width}{A number to define the width related to y axis.}
}
\value{
A ggplot
}
\description{
Display boxplot chart based of a summary.
}
