% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling-trainingtrips.R
\docType{data}
\name{TrainingTrips}
\alias{TrainingTrips}
\title{TrainingTrips}
\format{
A data.table with variables of the original data.
}
\source{
TLC Trip Record Data
}
\usage{
TrainingTrips
}
\description{
Random 5\% sample of taxi trips that took place in Manhattan, Brooklyn and Queens during 2022.
}
\keyword{datasets}
