% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda-functions-zonecodesarcgis.R
\docType{data}
\name{ZoneCodesArcgis}
\alias{ZoneCodesArcgis}
\title{ZoneCodesArcgis}
\format{
A data frame of 7 variables:
\describe{
\item{ LocationID }{ An unique id realted to each location}
\item{ Borough }{ Borough for the location }
\item{ Zone }{ Zone for the location }
\item{ service_zone }{ Define the zone type }
\item{ Address }{ A complete address for each location }
\item{ lat }{ A latitute estimation based on Address }
\item{ long }{ A longitute estimation based on Address }
}
}
\source{
TLC Trip Record Data and Arcgis
}
\usage{
ZoneCodesArcgis
}
\description{
Describe the locations used in the main data.
}
\keyword{datasets}
