% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_take_current_trip.R
\name{add_take_current_trip}
\alias{add_take_current_trip}
\title{Add and Evaluate Future Trip Alternatives}
\usage{
add_take_current_trip(
  trip_sample,
  point_mean_distance,
  parquet_path,
  future.scheduling = 1,
  future.chunk.size = NULL
)
}
\arguments{
\item{trip_sample}{A data frame containing the sample of trips to be evaluated. Each row represents a single trip.}

\item{point_mean_distance}{A data frame containing mean distances between different pickup and drop-off locations.
It should include at least PULocationID and DOLocationID columns.}

\item{parquet_path}{A string specifying the path to the Parquet file containing all historical trips data.}

\item{future.scheduling}{Average number of futures ("chunks") per worker.
If \code{0.0}, then a single future is used to process all elements
of \code{X}.
If \code{1.0} or \code{TRUE}, then one future per worker is used.
If \code{2.0}, then each worker will process two futures
(if there are enough elements in \code{X}).
If \code{Inf} or \code{FALSE}, then one future per element of
\code{X} is used.
Only used if \code{future.chunk.size} is \code{NULL}.}

\item{future.chunk.size}{The average number of elements per future ("chunk").
If \code{Inf}, then all elements are processed in a single future.
If \code{NULL}, then argument \code{future.scheduling} is used.}
}
\value{
A data table with the original trip sample and additional columns:
\itemize{
\item percentile_75_performance: The 75th percentile of performance for future trips.
\item take_current_trip: An integer (0 or 1) indicating whether the current trip should be taken based on performance metrics.
}
}
\description{
This function identifies and evaluates future trip alternatives for a given sample of trips.
It considers various constraints such as proximity to the current location, time window,
and trip characteristics like wheelchair accessibility, to determine whether to take the current trip.
}
\details{
The function connects to a DuckDB database to retrieve historical trip data. For each trip in the sample,
it fetches potential future trip alternatives that meet specific criteria:
\itemize{
\item The request time must be between 3 seconds and 15 minutes after the current trip's request time.
\item Future trips must be from the same company (hvfhs_license_num) as the current trip.
\item If the current trip is not for a wheelchair-accessible vehicle, future trips must also not be for a wheelchair-accessible vehicle.
}

Each valid future trip is evaluated based on the distance from the current trip and its potential profitability,
calculated as performance_per_hour (pay per hour, adjusted for waiting time). The function then compares the performance
of each trip against the 75th percentile of performance for future trips and decides whether to take the current trip.
}
