% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_zone_manually.R
\name{clean_zone_manually}
\alias{clean_zone_manually}
\title{Clean zone errors}
\usage{
clean_zone_manually(dt)
}
\arguments{
\item{dt}{A \code{data.table} with columns \code{lat}, \code{long}, \code{LocationID} and \code{Zone}.}
}
\value{
A data.table
}
\description{
The purpose of this function is correcting wrong locations like happened with:
\itemize{
\item Newark Airport
\item Governor's Island
\item Ellis Island
\item Liberty Island
}

It update the zone name to have a name for each location to avoid point overlapping.
}
