% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_pct.R
\name{count_pct}
\alias{count_pct}
\title{Counts rows based con columns provided}
\usage{
count_pct(x, ..., wt = NULL, sort = TRUE)
}
\arguments{
\item{x}{A lazy data frame arrow connection.}

\item{...}{Variables to group by}

\item{wt}{If a variable, computes sum(wt) for each group.}

\item{sort}{If TRUE, will show the largest groups at the top.}
}
\value{
A data.frame
}
\description{
It counts the number rows where each unique value repeated in the columns
selected arranging there results in descent order and adds a percentage
column after collecting the results from \code{arrow}.
}
\examples{
set.seed(1234)

arrow_con <- 
  data.frame(char1 = sample(LETTERS[1:3], 100, replace = TRUE),
             char2 = sample(LETTERS[4:6], 100, replace = TRUE)) |>
  arrow::arrow_table()

count_pct(arrow_con, char1)
count_pct(arrow_con, char2)
count_pct(arrow_con, char1, char2)
}
