% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_parquet.R
\name{sample_parquet}
\alias{sample_parquet}
\title{Take a sample from a parquet file}
\usage{
sample_parquet(file_path, valid_zones = NULL, prob = 0.05, seed = 1)
}
\arguments{
\item{file_path}{A character file name or URI.}

\item{valid_zones}{A vector of zones of the trips you want to sample from.}

\item{prob}{The proportion of rows to keep in the sample.}

\item{seed}{Number to keep the sample reprodicible.}
}
\value{
A data.table
}
\description{
Take a sample from a parquet file
}
\examples{
df <- data.frame(x = 1:100)
file_path <- tempfile(fileext = ".parquet")

arrow::write_parquet(df, file_path)

sample_parquet(file_path)
}
