% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trips.R
\name{simulate_trips}
\alias{simulate_trips}
\title{Simulates a day of work}
\usage{
simulate_trips(
  df,
  start_datetime,
  start_zone,
  minutes_next_trip,
  end_datetime,
  end_zones,
  closest_zone,
  borough_zones,
  seed_num = 1234
)
}
\arguments{
\item{df}{a data.frame or arrow connection with trips to sample from}

\item{start_datetime}{a date with an associated time when the taxi driver start to work}

\item{start_zone}{a number to select to the starting zone}

\item{minutes_next_trip}{an integer to define the limit time needed before extending the filters}

\item{end_datetime}{a date with an associated time when the taxi driver stop looking for new trips}

\item{end_zones}{a vector of number defining all possible zones to drive}

\item{closest_zone}{a named vector pointing the closest zone from the taxi driver is waiting for a need trip in order to start a trip}

\item{borough_zones}{a data.frame with the integer column LocationID with all possible zone ids and the list column id_list with contain all other zones ids related to the LocationID's borough}

\item{seed_num}{a single value, interpreted as an integer to make possible to reproduce random events}
}
\value{
A data.frame
}
\description{
This function sample from a data.frame or arrow connection all the trips needed
to complete the a working day form taxi driver.
}
