% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trips.R
\name{simulate_trips}
\alias{simulate_trips}
\title{Simulate a date of work for a single taxi driver.}
\usage{
simulate_trips(conn, start_points, model = NULL)
}
\arguments{
\item{conn}{Duckdb connection to NycTrips and PointMeanDistance tables.}

\item{start_points}{A data.table with the initial trips of each simulation.}

\item{model}{A tidymodels workflow.}
}
\value{
A data.table.
}
\description{
Simulate a date of work for a single taxi driver.
}
